Changes Metadata In TGEvent To JsonElement To Prevent GSON Exception
If the metadata field is a String, GSON can handle the deserialzation. However if the metadata field contains a Json structure, GSON cannot handle deserialzation to a simple string and instead tries to process the json structure. This changes the type from String to JsonElement in order to keep the same flexibility and leave it up to the consumer to implement the JsonElement deserialzatio/processing logic.